Create cancelWindow.java
So I completed the GUI for cancelling appoitment. We would need to make a for loop and possibly another arrayList to update the values in the list. We'll also need to have a variable to store the index we'll be deleting from the arrayList once the cancel button is clicked. The cancel button needs an action listener to delete, currently it's jsut a button without an action listener. Furthermore, I renamed some of the variable names to make it easier to view which variable is which in the interface.